Added function to remove schedule on state change
In my workflow, I wish to remove the schedule from tasks when they are
switched to a designated waiting state. To do this, I use a function
that will unshcedule the task upon state change. This function uses
a variable that contains a list of states it should act upon.
The function is activated by adding it to the org-trigger-hook.